Move functions from memefish.go
